+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 -   Snapshot formatting check to CI using `djlint`
 -   Autoformatting of snapshots using `djlint`
 -   Testing across Django versions 2.2 - 4.0 and Python versions 3.8 - 3.11 using `tox`
+-   Support for dividers on checkbox fields
 
 ### Changed
 
@@ -29,3 +30,4 @@
 ### Fixed
 
 -   `Field.select` label size and tag can be changed.
+-   `DateInputField` raises a `ValidationError` (instead of `ValueError`) when given invalid input